% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catMat.R
\name{catMat}
\alias{catMat}
\title{Convert categorical cell state vector into categorical matrix.}
\usage{
catMat(
  cell_states,
  num_states = NULL,
  state_order = NULL,
  sparse = TRUE,
  unformatted = FALSE
)
}
\arguments{
\item{cell_states}{A vector of cell states from a phylogeny.}

\item{num_states}{(optional) number of possible cell states.}

\item{state_order}{(optional) A vector of cell state names in order.}

\item{sparse}{(default=TRUE) returns sparse state matrix.}
}
\description{
Convert categorical cell state vector into categorical matrix.
}
