% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euc.R
\name{constrain_prolif}
\alias{constrain_prolif}
\title{Constrain cell state-specific proliferation rates using total proliferation}
\usage{
constrain_prolif(random_values, total_prolif, state_freqs)
}
\arguments{
\item{random_values}{Guess.}

\item{total_prolif}{Total proliferation rate.}

\item{state_freqs}{Vector cell state frequencies.}
}
\description{
Constrain cell state-specific proliferation rates using total proliferation
}
