% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrectP.R
\name{CorrectP}
\alias{CorrectP}
\title{CorrectP}
\usage{
CorrectP(P)
}
\arguments{
\item{P}{Square matrix.}
}
\value{
Pc Transition probability matrix, derived by correcting elements of \code{P}.
}
\description{
This function normalizes rows of the transition probability matrix P to sum to 1.
}
