% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euc.R
\name{euc}
\alias{euc}
\title{Euclidean distance}
\usage{
euc(x, y)
}
\arguments{
\item{x, y}{Matrices}
}
\value{
Euclidean distance between \code{x} and \code{y}.
}
\description{
Computed Euclidean distance between x and y.
}
