% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euc.R
\name{euc_reg}
\alias{euc_reg}
\title{Euclidead distance with regulaizer}
\usage{
euc_reg(x, y, a)
}
\arguments{
\item{x}{Matrix, vector, or scalar.}

\item{y}{Matrix, vector, or scalar with same dimensions as x.}

\item{a}{Matrix of parameters to regularize Euclidean distances for optimization.}
}
\description{
Euclidead distance with regulaizer
}
