% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp.tree.dist.R
\name{exp.tree.dist}
\alias{exp.tree.dist}
\title{Get the (inverse) exponential tree distances between all cells}
\usage{
exp.tree.dist(tree, node=TRUE, norm=TRUE)
}
\arguments{
\item{tree}{A phylogeny.}

\item{node}{Logical. Default TRUE. Tree distances are measured in terms of nodes. If set to FALSE distances in terms of branch lengths.}

\item{norm.}{Logical. Default TRUE. If TRUE, iverse tree distances are row normalized before sum normalizing.}
}
\value{
A matrix of inverse phylogenetic distances between cells.
}
\description{
This function gets the (inverse) exponential tree or phylogenetic distances, in terms of node distances or branch lengths, between all cell pairs. Exponential tree distances are normalized to sum to one.
}
