% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHpro.R
\name{fitPATHpro}
\alias{fitPATHpro}
\title{Fit PATHpro by optimizing \code{PATHpro_loss()}}
\usage{
fitPATHpro(
  initial_parameter_guess,
  pair_freq_list,
  state_freqs,
  pair_pat_dist = 2,
  total_prolif_rate = NULL,
  cell_prolif_rates = NULL,
  alph = 0,
  xtol = 1e-15,
  rtol = 1e-15,
  atol = 1e-15
)
}
\arguments{
\item{initial_parameter_guess}{Vector of paameter value guesses.}

\item{pair_freq_list}{List of cell state pair frequencies from phylogeny.}

\item{state_freqs}{Vector of cell state frequencies.}

\item{pair_pat_dist}{Vector of times/patristic distances corresponding to pair_freq_list.}

\item{total_prolif_rate}{Total proliferation rate.}

\item{cell_prolif_rates}{Vector of cell state proliferation rates.}

\item{alph}{Regulizer penalty weight to use when comparing pair_freq_list with those using guessed parameters.}

\item{xtol, rtol, atol}{Tolerances used for optimization using \code{nlminb()}.}
}
\value{
A list containing:\tabular{ll}{
\code{out} \tab Minimal sum of Euclidean distances and regulizer penalty for input and inference-based state pair frequencies.  \cr
\code{Q} \tab Inferred transition rate matrix. \cr
\code{P} \tab Inferred transition probability matrix. \cr
\code{gamma} \tab Inferred cell state specific proliferation rates (if not provided).
}
}
\description{
Fit PATHpro by optimizing \code{PATHpro_loss()}
}
