% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idealized_phylogeny.R
\name{idealized_phylogeny}
\alias{idealized_phylogeny}
\title{Simulate an idealized phylogeny}
\usage{
idealized_phylogeny(g, Q)
}
\arguments{
\item{g}{Number of generations. Phylogeny will have 2^g cells.}

\item{Q}{Transition rate matrix.}
}
\description{
First, this function uses \code{pbtree} from phytools to generate an idealized phylogeny.
Here, an idealized phylogeny, is a discrete-time, completely sampled, bifurcating, and perfectly balanced phylogeny.
Second, cell state transitions are simulated as a Markov chain with transition rate matrix Q.
}
