% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv.tree.dist.R
\name{inv.tree.dist}
\alias{inv.tree.dist}
\title{Get the inverse tree distances between all cells}
\usage{
inv.tree.dist(tree, node = TRUE, norm = TRUE)
}
\arguments{
\item{tree}{A phylogeny.}

\item{node}{Logical. Default TRUE. Tree distances are measured in terms of nodes. If set to FALSE distances in terms of branch lengths.}

\item{norm.}{Logical. Default TRUE. If TRUE, iverse tree distances are row normalized before sum normalizing.}
}
\value{
A matrix of inverse phylogenetic distances between cells.
}
\description{
This function gets the inverse tree or phylogenetic distances, in terms of node distances or branch lengths, between all cell pairs. Inverse tree distances are normalized to sum to one.
}
