% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euc.R
\name{makeQ}
\alias{makeQ}
\title{Make Q matrix from vector of parameters}
\usage{
makeQ(parameters, dimensions)
}
\arguments{
\item{parameters}{Vector of parameters.}

\item{dimensions}{Dimension of square output matrix.}
}
\value{
Transition rate matrix.
}
\description{
Make Q matrix from vector of parameters
}
