% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize vector to sum to 1.}
\usage{
normalize(x)
}
\arguments{
\item{x}{input vector with all positive elements.}
}
\value{
normalized vector.
}
\description{
Normalize vector to sum to 1.
}
