% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_node_depth.R
\name{one_node_depth}
\alias{one_node_depth}
\title{Get weight matrix of cells that share a most recent common ancestor at a node depth of 1.}
\usage{
one_node_depth(tree)
}
\arguments{
\item{tree}{A phylogeny.}
}
\value{
A list containing:\tabular{ll}{
\code{W} \tab weight matrix. \cr
\code{mean.pat} \tab mean patristic distance separating cell pairs. \cr
}
}
\description{
Get weight matrix of cells that share a most recent common ancestor at a node depth of 1.
}
