% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_weight_matrices.R
\name{one_node_tree_dist}
\alias{one_node_tree_dist}
\title{Phylogenetic weight matrix for a node-depth of one.}
\usage{
one_node_tree_dist(tree, norm = TRUE)
}
\arguments{
\item{tree}{A phylogeny.}

\item{norm}{Logical. Default TRUE. If true, the phylogenetic weight matrix is first row normalized before it is sum normalized.}
}
\description{
This function computes the sum normalized phylogenetic weight matrix
using the node-depth of one only weighting function. This function
returns an N by N dimensional matrix with a value in the ijth position
if cell i and j are separated by one node on the phylogeny and a 0 otherwise.
}
