% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATH.inference.R
\name{op_Pinf}
\alias{op_Pinf}
\title{Infer transition probability matrix using optimization.}
\usage{
op_Pinf(n, f, p)
}
\arguments{
\item{n}{number of cell states}

\item{f}{State pair frequencies from phylogeny.}

\item{p}{Mean patristic distances between cell pairs.}
}
\value{
PATH inference.
}
\description{
Infer transition probability matrix using optimization.
}
