% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATH.inference.R
\name{P_inf}
\alias{P_inf}
\title{PATH inference given cell state vector, phylogenetic weight matrix, and mean patristic distances}
\usage{
P_inf(state_vector, weight_matrix, mspd)
}
\arguments{
\item{state_vector}{Vector of cell states.}

\item{weight_matrix}{Phylogenetic weight matrix.}

\item{mspd}{Mean patristic distances between cells represented in weight_matrix.}
}
\description{
PATH inference given cell state vector, phylogenetic weight matrix, and mean patristic distances
}
