% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATH.inference.R
\name{P_inf.XW}
\alias{P_inf.XW}
\title{PATH inference for given patristic distances, cell state, and weight matrices.}
\usage{
P_inf.XW(X, W, mspd)
}
\arguments{
\item{X}{Cell state matrix.}

\item{W}{Phylogenetic weight matrix for specific depth.}

\item{mspd}{Mean patristic distances between cells.}
}
\description{
PATH inference for given patristic distances, cell state, and weight matrices.
}
