% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parM.R
\name{parM}
\alias{parM}
\title{Calculate heritability in parallel}
\usage{
parM(d, w, break.length = 100, core_num = 2)
}
\arguments{
\item{d}{Gene data matrix.}

\item{w}{Phylogenetic weight matrix.}

\item{break.length}{Number of genes to measure at a time.}

\item{core_num}{Number of cores for parallel processing.}
}
\description{
Calculate heritability in parallel
}
