% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATH.inference.R
\name{PATH_inf}
\alias{PATH_inf}
\title{Infer cell state transition probabilities from a phylogeny.}
\usage{
PATH_inf(
  tree,
  cell_states = "states",
  nstates = NULL,
  impute_branches = FALSE,
  sample_rate_est = NULL,
  birth = 1,
  death = 0
)
}
\arguments{
\item{tree}{A phylogeny.}

\item{cell_states}{Location of cell state vector.}

\item{nstates}{Number of possible cell states.}

\item{impute_branches}{Logical. Default FALSE. Use and impute phylogenetic branch lengths.}

\item{sample_rate_est}{Default NULL. Used for branch length imputation if impute_branches=TRUE.}

\item{birth}{Default 1. Used if impute_branches=TRUE.}

\item{death}{Default 0. Used if impute_branches=TRUE.}
}
\value{
A list containing:\tabular{ll}{
\code{P} \tab PATH inference of transition probability matrix P at time t=1. \cr
\code{Pt} \tab PATH inference of P for time proportional to the mean branch length distance at a node depth 1. \cr
\code{Q} \tab PATH inference of transition rate matrix Q = logm(P).
}
}
\description{
This function infers cell state transitions from the distribution
of states on a phylogeny.
}
