% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHpro.R
\name{PATHpro}
\alias{PATHpro}
\title{PATHpro}
\usage{
PATHpro(
  tree,
  z,
  depth = 1,
  prolif = NULL,
  total_prolif = NULL,
  guess_list = NULL,
  alpha = NULL
)
}
\arguments{
\item{tree}{Phylogeny.}

\item{z}{Cell state matrix.}

\item{depth}{Max node depth to use for computing phylogenetic correlations.}

\item{prolif}{Vector of cell state proliferation rates.}

\item{total_prolif}{Total proliferation rate.}

\item{guess_list}{List of initial parameter guess vectors.}

\item{alpha}{Regulaizer penalty weight.}
}
\description{
PATHpro
}
