% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHpro.R
\name{PATHpro.XW}
\alias{PATHpro.XW}
\title{PATHpro inference using input cell state and weight matrices.}
\usage{
PATHpro.XW(
  X,
  W,
  t,
  guess_list = NULL,
  total_prolif_rate_est = NULL,
  cell_prolifs = NULL,
  alpha0 = 0
)
}
\arguments{
\item{X}{Cell state matrix.}

\item{W}{Phylogenetic weight matrix, or list of weight matrices.}

\item{t}{Mean pairwise patristic distances corresponding to \code{W}.}

\item{guess_list}{List of initial parameter guesses (optional).}

\item{total_prolif_rate_est}{Estimate of total proliferation rate (optional).}

\item{cell_prolifs}{Vector of cell state-specific prolfieration rates, if known (optional).}

\item{alpha0}{Regulizer penalty weight.}
}
\value{
A list containing:\tabular{ll}{
\code{out} \tab Minimal sum of Euclidean distances and regulizer penalty for input and inference-based state pair frequencies.  \cr
\code{Q} \tab Inferred transition rate matrix, \eqn{\mathbf{Q}}. \cr
\code{P} \tab Inferred transition probability matrix, \eqn{\mathbf{P} = e^{\mathbf{Q}}}. \cr
\code{gamma} \tab Inferred cell state-specific proliferation rates (if not provided to \code{cell_prolifs}), \eqn{\mathbf{\gamma}}.
}
}
\description{
PATHpro inference using input cell state and weight matrices.
}
\details{
\eqn{\sum_{1}^{\text{depth}} \lVert \mathbb{P}^\top \mathbf{A} \mathbb{P} - \mathbf{X}^\top \overline{\mathbf{W}} \mathbf{X} \rVert + \alpha \sum_{i \neq j} Q_{ij}^2}
}
