% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHpro.R
\name{PATHpro.ZW}
\alias{PATHpro.ZW}
\title{PATHpro inference using input cell state and weight matrices.}
\usage{
PATHpro.ZW(
  z,
  w,
  t,
  guess_list = NULL,
  total_prolif_rate_est = NULL,
  cell_prolifs = NULL,
  cf0 = euc_reg
)
}
\arguments{
\item{z}{Cell state matrix.}

\item{w}{Phylogenetic weight matrix.}

\item{guess_list}{Vector of initial parameter guesses.}

\item{total_prolif_rate_est}{Estimate of total proliferation rate.}

\item{cell_prolifs}{Vector of cell state specific prolfieration rates.}

\item{cf0}{Cost function.}
}
\description{
PATHpro inference using input cell state and weight matrices.
}
