% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rP.R
\name{rP}
\alias{rP}
\title{Generate random transition matrix}
\usage{
rP(n, max.r = 0.1)
}
\arguments{
\item{n}{Number of distinct states (dimension of transition probability matrix)}

\item{max.r}{Parameter used for randomly generating off-diagonal entries of $Q$. Entries chosen from uniform distribution between 0 and max.r.}
}
\description{
\code{rQ} generates a random transition rate matrix $Q$.
\code{rP} generates a random transition probability matrix $P$.
}
