% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state2mat.sparse.R
\name{state2mat.sparse}
\alias{state2mat.sparse}
\title{Convert vector of cell states into matrix to compute phylogenetic correlations and perform transition inferences.}
\usage{
state2mat.sparse(states, n = NULL)
}
\arguments{
\item{states}{A vector of cell states from a phylogeny.}

\item{n}{Number of possible cell states.}
}
\description{
Convert vector of cell states into matrix to compute phylogenetic correlations and perform transition inferences.
}
