% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcor.R
\name{xcor}
\alias{xcor}
\title{Compute cross-correlations.}
\usage{
xcor(data, weight.matrix)
}
\value{
A list containing:\tabular{ll}{
\code{Morans.I} \tab Bivariate cross-correlations. Diagonal elements are auto-correlations. \cr
\code{Z.scores} \tab A matrix of analytical z scores for cross-correlations. \cr
\code{Expected.I} \tab Expected cross-correlations for a permutation. \cr
\code{Var.I} \tab Cross-correlation variance under permutations. \cr
\code{one.sided.pvalue} \tab One-sided p-values derived from z scores. \cr
}
}
\description{
This function computes both auto- and cross-correlations for
arguments: "data", an N by n dimensional matrix, and
"weight.matrix", an N by N dimensional matrix.
In the context of computing phylogenetic correlations,
N is the number of cells (terminal phylogenetic nodes or leaves),
and n is the number of observations (e.g., RNA counts for n genes).
}
\examples{
# phy is a phylogenetic tree
X <- phy$data
W <- one_node.tree.dist(phy)
xcor(X, W)
}
