% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcor_gsea.R
\name{xcor_gsea}
\alias{xcor_gsea}
\title{Gene Set Enrichment Analysis (GSEA) of genes ranked by phylogenetic auto-correlation.}
\usage{
xcor_gsea(
  gene.data,
  weight.matrix,
  pathways,
  maxS = Inf,
  minS = 5,
  nperm = 10000
)
}
\arguments{
\item{gene.data}{Matrix of single-cell gene expression data. Each row represents an individual cell and each column a gene.}

\item{weight.matrix}{A normalized phylogenetic weight matrix.}

\item{pathways}{Pathways to test for enrichment.}
}
\description{
Gene Set Enrichment Analysis (GSEA) of genes ranked by phylogenetic auto-correlation.
}
