% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHpro.R
\name{fitPATHpro}
\alias{fitPATHpro}
\title{Fit PATHpro by optimizing \code{PATHpro_loss()}}
\usage{
fitPATHpro(
  initial_parameter_guess,
  xcor_list,
  state_freqs,
  pair_pat_dist = 2,
  total_prolif_rate = NULL,
  cell_prolif_rates = NULL,
  cf = euc_reg,
  xtol = 1e-15,
  rtol = 1e-15,
  atol = 1e-15
)
}
\arguments{
\item{initial_parameter_guess}{Vector of paameter value guesses.}

\item{xcor_list}{List of cell state pair frequencies from phylogeny.}

\item{state_freqs}{Vector of cell state frequencies.}

\item{pair_pat_dist}{Vector of times/patristic distances corresponding to xcor_list.}

\item{total_prolif_rate}{Total proliferation rate.}

\item{cell_prolif_rates}{Vector of cell state proliferation rates.}

\item{cf}{Cost function to use when comparing xcor_list with those using guessed parameters.}
}
\description{
Fit PATHpro by optimizing \code{PATHpro_loss()}
}
