% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHpro.R
\name{PATHpro_loss}
\alias{PATHpro_loss}
\title{PATHpro loss function for optimization}
\usage{
PATHpro_loss(
  guess,
  pair_freqs,
  times,
  state_freqs,
  total_prolif,
  prolif_rates,
  alpha = 0
)
}
\arguments{
\item{guess}{Vector of parameter value guesses.}

\item{pair_freqs}{List of cell state pair frequencies from phylogeny.}

\item{times}{Vector of times/patristic distances corresponding to pair_freqs.}

\item{state_freqs}{Vector of cell state frequencies.}

\item{total_prolif}{Total proliferation rate.}

\item{prolif_rates}{Vector of cell state proliferation rates.}

\item{alpha}{Regulizer penalty weight to use when comparing pair_freqs with those using guessed parameters.}
}
\description{
PATHpro loss function for optimization
}
