% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHpro.R
\name{PATHpro_loss}
\alias{PATHpro_loss}
\title{PATHpro loss function for optimization}
\usage{
PATHpro_loss(guess, xcors, times, freqs, total_prolif, prolif_rates, cost_func)
}
\arguments{
\item{guess}{Vector of parameter value guesses.}

\item{xcors}{List of cell state pair frequencies from phylogeny.}

\item{times}{Vector of times/patristic distances corresponding to xcors.}

\item{freqs}{Vector of cell state frequencies.}

\item{total_prolif}{Total proliferation rate.}

\item{prolif_rates}{Vector of cell state proliferation rates.}

\item{cost_func}{Cost function to use when comparing xcors with those using guessed parameters.}
}
\description{
PATHpro loss function for optimization
}
