% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcor.R
\name{xcor}
\alias{xcor}
\title{Compute cross-correlations.}
\usage{
xcor(data, weight.matrix)
}
\arguments{
\item{data}{\eqn{N \times n} matrix of trait/state data, where \eqn{N} is number of cells (phylogeny leaves) and \eqn{n} is number of traits.}

\item{weight.matrix}{\eqn{N \times N} phylogenetic weight matrix.}
}
\value{
A list containing:\tabular{ll}{
\code{phy_cor} \tab Phylogenetic cross-correlations. Diagonal elements are phylogenetic auto-correlations. \cr
\code{Z.score} \tab A matrix of analytical z scores for cross-correlations. \cr
\code{one.sided.pvalue} \tab One-sided p-values derived from z scores. \cr
}
}
\description{
This function computes both auto- and cross-correlations for
arguments: \code{data}, an \eqn{N \times n} dimensional matrix, and
\code{weight.matrix}, an \eqn{N \times N} dimensional matrix.
In the context of computing phylogenetic correlations,
\eqn{N} is the number of cells (terminal phylogenetic nodes or leaves),
and \eqn{n} is the number of observations (e.g., RNA counts for \eqn{n} genes).
}
\examples{
phy <- ran_tree_with_states(matrix(c(-0.1, 0.1, 0.1, -0.1), 2, 2))
X <- catMat(phy$states)
W <- one_node_tree_dist(phy)
xcor(X, W)
}
